requirements.txt
httplib2
Django==1.8
requests
virtualenv
virtualenvwrapper
django-ipware
django_compressor
